+++ e2e/runner/solana.go
@@ -129,6 +129,7 @@ func (r *E2ERunner) CreateDepositSPLInstruction(
 	to solana.PublicKey,
 	receiver ethcommon.Address,
 	data []byte,
+	revertOptions *solanacontract.RevertOptions,
 ) solana.Instruction {
 	var err error
 	var depositSPLData []byte
@@ -137,6 +138,7 @@ func (r *E2ERunner) CreateDepositSPLInstruction(
 			Discriminator: solanacontract.DiscriminatorDepositSPL,
 			Amount:        amount,
 			Receiver:      receiver,
+			RevertOptions: revertOptions,
 		})
 		require.NoError(r, err)
 	} else {
@@ -145,6 +147,7 @@ func (r *E2ERunner) CreateDepositSPLInstruction(
 			Amount:        amount,
 			Receiver:      receiver,
 			Memo:          data,
+			RevertOptions: revertOptions,
 		})
 		require.NoError(r, err)
 	}
@@ -242,6 +245,7 @@ func (r *E2ERunner) SPLDepositAndCall(
 	mintAccount solana.PublicKey,
 	receiver ethcommon.Address,
 	data []byte,
+	revertOptions *solanacontract.RevertOptions,
 ) solana.Signature {
 	// ata for pda
 	pda := r.ComputePdaAddress()
@@ -264,9 +268,10 @@ func (r *E2ERunner) SPLDepositAndCall(
 		pdaAta,
 		receiver,
 		data,
+		revertOptions,
 	)
 
-	limit := computebudget.NewSetComputeUnitLimitInstruction(100000).Build() // 100k compute unit limit
+	limit := computebudget.NewSetComputeUnitLimitInstruction(500000).Build() // 500k compute unit limit
 	feesInit := computebudget.NewSetComputeUnitPriceInstructionBuilder().
 		SetMicroLamports(100000).Build() // 0.1 lamports per compute unit
 	signedTx := r.CreateSignedTransaction(
@@ -455,7 +460,7 @@ func (r *E2ERunner) SOLDepositAndCall(
 	instruction := r.CreateDepositInstruction(signerPrivKey.PublicKey(), receiver, data, amount.Uint64(), revertOptions)
 
 	// create and sign the transaction
-	limit := computebudget.NewSetComputeUnitLimitInstruction(100000).Build() // 100k compute unit limit
+	limit := computebudget.NewSetComputeUnitLimitInstruction(500000).Build() // 500k compute unit limit
 	feesInit := computebudget.NewSetComputeUnitPriceInstructionBuilder().
 		SetMicroLamports(100000).Build() // 0.1 lamports per compute unit
 	signedTx := r.CreateSignedTransaction(
@@ -549,10 +554,7 @@ func (r *E2ERunner) WithdrawAndCallSOLZRC20(
 	utils.RequireTxSuccessful(r, receipt, "approve")
 
 	// create encoded msg
-	connected := solana.MustPublicKeyFromBase58("4xEw862A2SEwMjofPkUyd4NEekmVJKJsdHkK3UkAtDrc")
-	connectedPda, err := solanacontract.ComputeConnectedPdaAddress(connected)
-	require.NoError(r, err)
-	abiArgs, err := solanacontract.GetExecuteMsgAbi()
+	connectedPda, err := solanacontract.ComputeConnectedPdaAddress(ConnectedProgramID)
 	require.NoError(r, err)
 	msg := solanacontract.ExecuteMsg{
 		Accounts: []solanacontract.AccountMeta{
@@ -564,7 +566,7 @@ func (r *E2ERunner) WithdrawAndCallSOLZRC20(
 		Data: data,
 	}
 
-	msgEncoded, err := abiArgs.Pack(msg)
+	msgEncoded, err := msg.Encode()
 	require.NoError(r, err)
 
 	// withdraw
@@ -639,8 +641,6 @@ func (r *E2ERunner) WithdrawAndCallSPLZRC20(
 	connectedPdaAta := r.ResolveSolanaATA(r.GetSolanaPrivKey(), connectedPda, r.SPLAddr)
 	randomWalletAta := r.ResolveSolanaATA(r.GetSolanaPrivKey(), r.GetSolanaPrivKey().PublicKey(), r.SPLAddr)
 
-	abiArgs, err := solanacontract.GetExecuteMsgAbi()
-	require.NoError(r, err)
 	msg := solanacontract.ExecuteMsg{
 		Accounts: []solanacontract.AccountMeta{
 			{PublicKey: [32]byte(connectedPda.Bytes()), IsWritable: true},
@@ -655,7 +655,7 @@ func (r *E2ERunner) WithdrawAndCallSPLZRC20(
 		Data: data,
 	}
 
-	msgEncoded, err := abiArgs.Pack(msg)
+	msgEncoded, err := msg.Encode()
 	require.NoError(r, err)
 
 	// withdraw