+++ changelog.md
@@ -18,6 +18,8 @@
 * [3756](https://github.com/zeta-chain/node/pull/3756) - parse revert options in solana inbounds
 * [3765](https://github.com/zeta-chain/node/pull/3765) - support cancelling Sui rejected withdrawal
 * [3792](https://github.com/zeta-chain/node/pull/3792) - add compliance check for Sui inbound and outbound
+* [3790](https://github.com/zeta-chain/node/pull/3790) - integrate execute revert
+* [3797](https://github.com/zeta-chain/node/pull/3797) - integrate execute SPL revert
 * [3793](https://github.com/zeta-chain/node/pull/3793) - support Sui withdrawAndCall using the PTB transaction
 
 ### Refactor